requirements.txt
asgiref==3.8.1
Django==5.0.4
packaging==24.0
pillow==10.3.0
plotly==5.21.0
sqlparse==0.5.0
tenacity==8.2.3
django-environ==0.8.1